+++ lib/csv_plus_plus/language/entities/cell_reference.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative '../ast_builder'
 require_relative './entity'
 
 module CSVPlusPlus
@@ -10,6 +11,18 @@ module CSVPlusPlus
       # @attr_reader cell_reference [String] The cell reference in A1 format
       class CellReference < Entity
         attr_reader :cell_reference
+
+        # Create a +CellReference+ to the given indexes
+        #
+        # @param cell_index [Integer] The current cell index
+        # @param row_index [Integer] The current row index
+        #
+        # @return [CellReference]
+        def self.from_index(cell_index:, row_index:)
+          return unless row_index || cell_index
+
+          new(::CSVPlusPlus::Language::ASTBuilder.ref(cell_index:, row_index:))
+        end
 
         # @param cell_reference [String] The cell reference in A1 format
         def initialize(cell_reference)