+++ spec/csv_plus_plus/language/can_define_references_spec.rb
@@ -1,18 +1,19 @@
 # frozen_string_literal: true
 
-describe ::CSVPlusPlus::CodeSection do
+describe ::CSVPlusPlus::Language::CanDefineReferences do
   let(:functions) { {} }
   let(:variables) { {} }
-  let(:code_section) { build(:code_section, variables:, functions:) }
+
+  let(:scope) { build(:scope, variables:, functions:) }
 
   describe '#def_variable' do
     let(:id) { 'foo' }
     let(:value) { build(:number_one) }
 
-    before { code_section.def_variable(id, value) }
+    before { scope.def_variable(id, value) }
 
     it 'sets the values in @variables' do
-      expect(code_section.variables).to(eq({ foo: build(:number_one) }))
+      expect(scope.variables).to(eq({ foo: build(:number_one) }))
     end
   end
 
@@ -20,16 +21,16 @@ describe ::CSVPlusPlus::CodeSection do
     let(:var_foo) { build(:variable_foo) }
     let(:id) { 'foo' }
 
-    before { code_section.def_variable(id, var_foo) }
+    before { scope.def_variable(id, var_foo) }
 
     it 'sets the function in @variables' do
-      expect(code_section.variables).to(eq({ foo: build(:variable_foo) }))
+      expect(scope.variables).to(eq({ foo: build(:variable_foo) }))
     end
 
     it 'overwrites previous definitions' do
       var_bar = build(:variable_bar)
-      code_section.def_variable(:foo, var_bar)
-      expect(code_section.variables).to(eq({ foo: build(:variable_bar) }))
+      scope.def_variable(:foo, var_bar)
+      expect(scope.variables).to(eq({ foo: build(:variable_bar) }))
     end
   end
 
@@ -37,16 +38,16 @@ describe ::CSVPlusPlus::CodeSection do
     let(:fn_foo) { build(:fn_foo) }
     let(:id) { 'foo' }
 
-    before { code_section.def_function(id, fn_foo) }
+    before { scope.def_function(id, fn_foo) }
 
     it 'sets the function in @functions' do
-      expect(code_section.functions).to(eq({ foo: build(:fn_foo) }))
+      expect(scope.functions).to(eq({ foo: build(:fn_foo) }))
     end
 
     it 'overwrites previous definitions' do
       fn_bar = build(:fn_bar)
-      code_section.def_function(:foo, fn_bar)
-      expect(code_section.functions).to(eq({ foo: build(:fn_bar) }))
+      scope.def_function(:foo, fn_bar)
+      expect(scope.functions).to(eq({ foo: build(:fn_bar) }))
     end
   end
 
@@ -54,7 +55,7 @@ describe ::CSVPlusPlus::CodeSection do
     let(:var_id) { :foo }
     let(:variables) { { foo: build(:number_one) } }
 
-    subject { code_section }
+    subject { scope }
 
     it { is_expected.to(be_defined_variable(var_id)) }
 
@@ -69,19 +70,13 @@ describe ::CSVPlusPlus::CodeSection do
     let(:fn_id) { :foo }
     let(:functions) { { foo: build(:fn_foo) } }
 
-    subject { code_section }
+    subject { scope }
 
     it { is_expected.to(be_defined_function(fn_id)) }
   end
 
-  describe '#to_s' do
-    subject { code_section.to_s }
-
-    it { is_expected.to(eq('CodeSection(functions: {}, variables: {})')) }
-  end
-
   describe '#verbose_summary' do
-    subject { code_section.verbose_summary }
+    subject { scope.verbose_summary }
 
     it { is_expected.to(match(/\#.*Variables/)) }
     it { is_expected.to(match(/\#.*Functions/)) }