+++ spec/csv_plus_plus/language/scope_spec.rb
@@ -4,86 +4,85 @@ describe ::CSVPlusPlus::Language::Scope do
   let(:runtime) { build(:runtime) }
   let(:scope) { described_class.new(runtime:) }
 
-  describe 'code_section=' do
-    let(:ast) do
-      build(:fn_call, name: :multiply, arguments: [build(:variable, id: :bar), build(:variable, id: :foo)])
-    end
-    let(:variables) do
-      {
-        bar: build(:number_one),
-        foo: build(:fn_call_add),
-        dep: ast
-      }
-    end
-    let(:functions) { {} }
-    let(:code_section) { build(:code_section, variables:, functions:) }
-
-    before { scope.code_section = code_section }
-
-    describe '@variables' do
-      subject { code_section.variables }
-
-      it 'resolves the variables in dep' do
-        expect(subject[:dep]).to(
-          eq(
-            build(
-              :fn_call,
-              name: :multiply,
-              arguments: [variables[:bar], variables[:foo]]
-            )
-          )
-        )
-      end
-
-      context 'with runtime variables' do
-        let(:ast) do
-          build(:fn_call, name: :multiply, arguments: [build(:variable, id: :rownum), build(:variable, id: :foo)])
-        end
-
-        it 'resolves just the static variables in dep' do
-          expect(subject[:dep]).to(
-            eq(
-              build(
-                :fn_call,
-                name: :multiply,
-                arguments: [
-                  build(:variable, id: :rownum),
-                  variables[:foo]
-                ]
-              )
-            )
-          )
-        end
-      end
-
-      context 'with undefined variables' do
-        let(:bad_variables) { { foo: build(:variable, id: :thisdoesnotexist) } }
-        let(:bad_code_section) { build(:code_section, variables: bad_variables) }
-
-        it 'should raise a SyntaxError' do
-          expect { scope.code_section = bad_code_section }
-            .to(raise_error(::CSVPlusPlus::Language::SyntaxError))
-        end
-      end
-    end
-
-    describe '@functions' do
-      subject { code_section.functions }
-
-      it 'resolves function dependencies' do
-        # TODO: not implemented yet
-      end
-
-      context 'with builtin functions' do
-        # TODO
-      end
-    end
-  end
+  #   describe 'code_section=' do
+  #     let(:ast) do
+  #       build(:fn_call, name: :multiply, arguments: [build(:variable, id: :bar), build(:variable, id: :foo)])
+  #     end
+  #     let(:variables) do
+  #       {
+  #         bar: build(:number_one),
+  #         foo: build(:fn_call_add),
+  #         dep: ast
+  #       }
+  #     end
+  #     let(:functions) { {} }
+  #     let(:code_section) { build(:code_section, variables:, functions:) }
+  #
+  #     before { scope.code_section = code_section }
+  #
+  #     describe '@variables' do
+  #       subject { code_section.variables }
+  #
+  #       it 'resolves the variables in dep' do
+  #         expect(subject[:dep]).to(
+  #           eq(
+  #             build(
+  #               :fn_call,
+  #               name: :multiply,
+  #               arguments: [variables[:bar], variables[:foo]]
+  #             )
+  #           )
+  #         )
+  #       end
+  #
+  #       context 'with runtime variables' do
+  #         let(:ast) do
+  #           build(:fn_call, name: :multiply, arguments: [build(:variable, id: :rownum), build(:variable, id: :foo)])
+  #         end
+  #
+  #         it 'resolves just the static variables in dep' do
+  #           expect(subject[:dep]).to(
+  #             eq(
+  #               build(
+  #                 :fn_call,
+  #                 name: :multiply,
+  #                 arguments: [
+  #                   build(:variable, id: :rownum),
+  #                   variables[:foo]
+  #                 ]
+  #               )
+  #             )
+  #           )
+  #         end
+  #       end
+  #
+  #       context 'with undefined variables' do
+  #         let(:bad_variables) { { foo: build(:variable, id: :thisdoesnotexist) } }
+  #         let(:bad_code_section) { build(:code_section, variables: bad_variables) }
+  #
+  #         it 'should raise a SyntaxError' do
+  #           expect { scope.code_section = bad_code_section }
+  #             .to(raise_error(::CSVPlusPlus::Language::SyntaxError))
+  #         end
+  #       end
+  #     end
+  #
+  #     describe '@functions' do
+  #       subject { code_section.functions }
+  #
+  #       it 'resolves function dependencies' do
+  #         # TODO: not implemented yet
+  #       end
+  #
+  #       context 'with builtin functions' do
+  #         # TODO
+  #       end
+  #     end
+  #   end
 
   describe '#resolve_cell_value' do
-    let(:code_section) { build(:code_section, functions:, variables: { foo: build(:number_one) }) }
+    let(:scope) { build(:scope, functions:, variables: { foo: build(:number_one) }, runtime:) }
     let(:functions) { {} }
-    let(:scope) { build(:scope, code_section:, runtime:) }
     let(:runtime) { build(:runtime, cell:) }
 
     let(:fn_call_celladjacent) { build(:fn_call, name: :celladjacent, arguments: [build(:cell_reference, ref: 'A')]) }
@@ -153,7 +152,7 @@ describe ::CSVPlusPlus::Language::Scope do
     subject { scope.to_s }
 
     it do
-      is_expected.to(match(/Scope\(code_section: , runtime: Runtime\(.*\)\)/))
+      is_expected.to(match(/Scope\(functions: .+, runtime: .+, variables: .+\)/))
     end
   end
 end